projects/identifier/requirements.txt
functions-framework==3.*
pillow==10.2.0
google-cloud-secret-manager==2.17.0
langchain-google-genai
langchain